- adding tasks in runtime

+++ app/src/main/java/com/example/android/todoapp_android/database/TaskDao.kt
@@ -10,7 +10,7 @@ interface TaskDao {
     fun getAll(): List<Task>
 
     @Query("SELECT * FROM tasks WHERE date LIKE:date")
-    fun loadAllByDate(date: Int?): List<Task>
+    fun loadAllByDate(date: Long?): List<Task>
 
 
     /*
